Load Settings sync panel on demand
