+++ tsconfig.json
@@ -236,6 +236,7 @@
     "js/settings-privacy.js",
     "js/settings-runtime.js",
     "js/settings-sync-panel.js",
+    "js/settings-sync-panel-impl.js",
     "js/settings-agent-access-panel.js",
     "js/shell-actions.js",
     "js/silhouette-paths.js",
